+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 
 <!---------------------------------->
+<a name="v0.0.63.0"></a>
+## v0.0.63.0
+
+### Panther Documentation
+- Updated to reflect changes in [PCIT-CPP v0.0.228.0](https://github.com/PCIT-Project/PCIT-CPP/blob/main/CHANGELOG.md#v0.0.228.0)
+- Added `-hel` / `-h` command-line arguments to generator
+- Added command line argument checking to generator
+
+
+<!---------------------------------->
 <a name="v0.0.62.0"></a>
 ## v0.0.62.0
 - Added `-quick-date` command-line argument to generator